Moved vt code to pkg and refactored it and its tests. Removed time wait for analysis
